Refactor futures data polling
