+++ pages/dashboard/markets.tsx
@@ -4,14 +4,13 @@ import { useTranslation } from 'react-i18next';
 import DashboardLayout from 'sections/dashboard/DashboardLayout';
 import Markets from 'sections/dashboard/Markets';
 import GitHashID from 'sections/shared/Layout/AppLayout/GitHashID';
-import { fetchMarkets } from 'state/futures/actions';
-import { usePollWhenReady } from 'state/hooks';
+import { usePollFuturesData } from 'state/futures/hooks';
 
 type MarketsProps = React.FC & { getLayout: (page: HTMLElement) => JSX.Element };
 
 const MarketsPage: MarketsProps = () => {
 	const { t } = useTranslation();
-	usePollWhenReady('fetchMarkets', fetchMarkets);
+	usePollFuturesData();
 
 	return (
 		<>
